// src/app/city-cast/interfaces/layer-properties.ts
export interface LayerProperties {
    layer_id?: string;
    source_id?: string;
    source_layer?: string;
    zoom?: any;
    color?: string;
    active_color?: string;
    is_using_data_joins?: boolean;
    max?: number;
    indexes?: any;
    networkIndexes?: any;
    networkMode?: string;
    max_line_width?: number;
}